Update tests to address review comments
